Library

Library

Public API

System Variables

Differentiation Methods

Sparsity

Constructors

Solvers

Calling Interface

Convenience Functions

Private API

Public API

System Variables

CoupledSystems.@var — Macro

@var var = val

Macro which modifies the expression var = val to var = NamedVar(:var, val)

julia> @var x = 5
  NamedVar{Int64}(:x, 5)

julia> x
  NamedVar{Int64}(:x, 5)

CoupledSystems.NamedVar — Type

NamedVar{T}

Defines a named variable

Field

name::Symbol: Variable name

value::T: Default variable value

CoupledSystems.name — Function

name(var::NamedVar)

Return the name of a named variable.

CoupledSystems.value — Function

value(var::NamedVar)

Return the value of a named variable.

value(val)

Return the value of a variable.

CoupledSystems.combine — Function

combine(vars::Tuple)

Combine the values of all variables in vars into a single vector

CoupledSystems.combine! — Function

combine!(y, vars::Tuple)

In-place version of combine

CoupledSystems.separate — Function

separate(vars::Tuple, values)

Return the values stored in values as separated variables rather than a single output vector. vars is a tuple of named variables corresponding to each returned parameter.

Differentiation Methods

CoupledSystems.AbstractDiffMethod — Type

AbstractDiffMethod

Abstract type representing different differentiation methods

CoupledSystems.AbstractAD — Type

AbstractAD

Abstract type representing different automatic differentation methods

CoupledSystems.ForwardAD — Type

ForwardAD{T} <: AbstractAD

Forward automatic differentiation using the ForwardDiff package.

Fields:

chunk::T: Chunk size

CoupledSystems.ReverseAD — Type

ReverseAD <: AbstractAD

Reverse automatic differentiation using the ReverseDiff package.

CoupledSystems.AbstractFD — Type

AbstractFD

Abstract type representing different finite difference methods

CoupledSystems.ForwardFD — Type

ForwardFD <: AbstractFD

Forward finite differencing as implemented by the FiniteDiff package

CoupledSystems.CentralFD — Type

CentralFD <: AbstractFD

Central finite differencing as implemented by the FiniteDiff package

CoupledSystems.ComplexFD — Type

ComplexFD <: AbstractFD

Complex step finite differencing as implemented by the FiniteDiff package

CoupledSystems.Forward — Type

Forward <: AbstractChainRuleMode

Type indicating that the chain rule should be applied in forward mode

CoupledSystems.Reverse — Type

Reverse <: AbstractChainRuleMode

Type indicating that the chain rule should be applied in reverse mode

CoupledSystems.Direct — Type

Direct <: AbstractSensitivityMode

Type indicating that the analytic sensitivity equations should be applied using the direct method.

CoupledSystems.Adjoint — Type

Adjoint <: AbstractSensitivityMode

Type indicating that the analytic sensitivity equations should be applied using the adjoint method.

Sparsity

CoupledSystems.AbstractSparsityPattern — Type

AbstractSparsityPattern

Abstract type representing the sparsity pattern of a jacobian matrix

CoupledSystems.DensePattern — Type

DensePattern <: AbstractSparsityPattern

Type which indicates that a jacobian is dense

CoupledSystems.SparsePattern — Type

SparsePattern{TI} <: AbstractSparsityPattern

Type which indicates that a jacobian is sparse and indicates the non-zero rows and columns

Fields:

rows::Vector{TI}: Row of each non-zero matrix element

cols::Vector{TI}: Column of each non-zero matrix element

Constructors

CoupledSystems.AbstractComponent — Type

AbstractComponent

Supertype for all components

CoupledSystems.AbstractExplicitComponent — Type

AbstractExplicitComponent <: AbstractComponent

Supertype for components defined by the vector valued output function y = f(x)

CoupledSystems.ExplicitComponent — Type

ExplicitComponent{TX, TY, TJ, TI, TO} <: AbstractExplicitComponent

System component defined by the explicit vector-valued output function: y = f(x)

Fields

f: In-place output function f(y, x)

df: In-place jacobian function df(dydx, x)

fdf: In-place combined output and jacobian function fdf(y, dydx, x)

x_f::TX: Inputs used to evaluate the outputs

x_df::TX: Inputs used to evaluate the jacobian

y::TY: Outputs

dydx::TJ: Jacobian

argin::TI: Tuple of named inputs to output function, defines size, type, and default values of inputs

argout::TO: Tuple of named outputs from output function, defines size and type of outputs

CoupledSystems.ExplicitSystem — Type

ExplicitSystem{TC, TX, TY, TJ, TD} <: AbstractExplicitComponent

Explicit system constructed from a chain of explicit system components called sequentially.

Fields

components::TC: Collection of components, in calling order

input_mapping::Vector{NTuple{2,Vector{Int}}}:

component_output_mapping::Vector{Vector{NTuple{2,Vector{Int}}}}:

component_input_mapping::Vector{Vector{NTuple{2,Int}}}:

output_mapping::Vector{NTuple{2,Int}}:

x_f::TX: Inputs used to evaluate the system outputs

x_df::TX: Inputs used to evaluate the system jacobian

y::TY: Storage for the system outputs

dydx::TJ: Storage for the system jacobian

argin::TI: Tuple of named inputs to output function, defines size, type, and default values of inputs

argout::TO: Tuple of named outputs from output function, defines size and type of outputs

mode::TD: Default direction in which to apply the chain rule (Forward() or Reverse())

CoupledSystems.AbstractImplicitComponent — Type

AbstractImplicitComponent <: AbstractComponent

Supertype for components defined by the vector valued residual function 0 = f(x, y)

CoupledSystems.ImplicitComponent — Type

ImplicitComponent{TX, TY, TR, TDRX, TDRY, TI, TO} <: AbstractImplicitComponent

System component defined by the vector-valued residual function: 0 = f(x, y)

Fields

f: In-place residual function f(r, x, y).

dfdx: In-place residual jacobian function with respect to the inputs dfdx(drdx, x, y)

dfdy: In-place residual jacobian function with respect to the outputs dfdy(drdy, x, y)

fdfdx: In-place combined residual and jacobian with respect to the inputs function fdfdx(r, drdx, x, y).

fdfdy: In-place combined residual and jacobian with respect to the outputs function fdfdy(r, drdy, x, y).

fdf: In-place combined residual and jacobians function fdf(r, drdx, drdy, x, y).

x_f::TX: x used to evaluate f

y_f::TY: y used to evaluate f

x_dfdx::TX: x used to evaluate dfdx

y_dfdx::TY: y used to evaluate dfdx

x_dfdy::TX: x used to evaluate dfdy

y_dfdy::TY: y used to evaluate dfdy

r::TR: cache for residual`

drdx::TDRX: cache for residual jacobian with respect to x

drdy::TDRY: cache for residual jacobian with respect to y

argin::TI: Tuple of named inputs to output function, defines size, type, and default values of inputs

argout::TO: Tuple of named outputs from output function, defines size, type, and default values of outputs

CoupledSystems.ImplicitSystem — Type

ImplicitSystem{TC, TX, TY, TR, TDRX, TDRY, TI, TO} <: AbstractImplicitComponent

Implicit system constructed from interdependent explicit and/or implicit system components.

Fields

components::TC: Collection of components, in preferred calling order

component_input_mapping::Vector{Vector{NTuple{2,Int}}}:

x_f::TX: x used to evaluate f

y_f::TY: y used to evaluate f

x_dfdx::TX: x used to evaluate dfdx

y_dfdx::TY: y used to evaluate dfdx

x_dfdy::TX: x used to evaluate dfdy

y_dfdy::TY: y used to evaluate dfdy

r::TR: cache for residual`

drdx::TDRX: cache for residual jacobian with respect to x

drdy::TDRY: cache for residual jacobian with respect to y

argin::TI: Tuple of named inputs to output function, defines size, type, and default values of inputs

argout::TO: Tuple of named outputs from output function, defines size, type, and default values of outputs

idx::Vector{Int}: Index used for accessing outputs/residuals for each component

Solvers

CoupledSystems.AbstractSolver — Type

AbstractSolver

Supertype for solvers that may be used to solve implicit systems of equations.

CoupledSystems.Newton — Type

Newton{T} <: AbstractSolver

Newton solver, as provided by the NLsolve package. Stores the keyword arguments which are passed to NLsolve in the field kwargs

Calling Interface

CoupledSystems.outputs — Function

outputs(component::AbstractComponent)

Return the outputs stored in component.

outputs(component::AbstractExplicitComponent, x)

Evaluate the outputs of an explicit system component.

This does not update any of the values stored in component

CoupledSystems.outputs! — Function

outputs!(component::AbstractExplicitComponent, y, x)

Evaluate the outputs of an explicit system component and store the result in y.

This does not update any of the values stored in component

outputs!(component::ExplicitSystem, y, xsub, ysub, x)

Evaluate the outputs of an explicit system and store the result in y. Store the intermediate results in xsub and ysub.

This does not update any of the values stored in component

outputs!(component::AbstractExplicitComponent, x)

Evaluate the outputs of an explicit system component.

Return the result and store in component.y.

CoupledSystems.outputs!! — Function

outputs!!(component::AbstractExplicitComponent, x)

Force (re-)evaluation of the outputs of an explicit system component.

Return the result and store in component.y.

CoupledSystems.outputs!!! — Function

outputs!!!(component::AbstractExplicitComponent, x)

Force (re-)evaluation of the outputs of an explicit system component. Also force (re-)evaluation of the outputs of the system's subcomponents.

Return the result and store in component.y.

CoupledSystems.jacobian — Function

jacobian(component::AbstractExplicitComponent)

Return the jacobian of the outputs with respect to the inputs stored in component

jacobian(component::AbstractExplicitComponent, x)

Evaluate the jacobian of the outputs with respect to the inputs.

This does not update any of the values stored in component

CoupledSystems.jacobian! — Function

jacobian!(component::AbstractExplicitComponent, dydx, x)

Evaluate the jacobian of the outputs with respect to the inputs and store the result in dydx

This does not update any of the values stored in component

jacobian!(component::ExplicitSystem, dydx, xsub, ysub, dsub, x, mode=component.mode)

Evaluate the jacobian of the outputs with respect to the inputs and store the result in dydx. Store the intermediate subcomponent x, y and dydx values in xsub, ysub, and dsub respectively.

This does not update any of the values stored in component

jacobian!(component::AbstractExplicitComponent, x)

Evaluate the jacobian of the outputs with respect to the inputs.

Return the result and store in component.dydx.

CoupledSystems.jacobian!! — Function

jacobian!!(component::AbstractExplicitComponent, x)

Force (re-)evaluation of the jacobian of the outputs with respect to the inputs.

Return the result and store in component.dydx.

CoupledSystems.jacobian!!! — Function

jacobian!!!(component::AbstractExplicitComponent, x)

Force (re-)evaluation of the jacobian of the outputs with respect to the inputs. Also force (re-)evaluation of the jacobians of the system's subcomponents.

Return the result and store in component.dydx.

CoupledSystems.outputs_and_jacobian — Function

outputs_and_jacobian(component::AbstractExplicitComponent)

Return the outputs and their derivatives with respect to the inputs stored in component.

outputs_and_jacobian(component::AbstractExplicitComponent, x)

Evaluate the outputs and their derivatives with respect to the inputs.

This does not update any of the values stored in component

CoupledSystems.outputs_and_jacobian! — Function

outputs_and_jacobian!(component::AbstractExplicitComponent, y, dydx, x)

Evaluate the outputs and their derivatives with respect to the inputs and store the results in y and dydx

This does not update any of the values stored in component

outputs_and_jacobian!(component::ExplicitSystem, y, dydx, xsub, ysub, dsub,
    x, mode=component.mode)

Evaluate the outputs and their derivatives with respect to the inputs and store the results in y and dydx. Store the intermediate subcomponent x, y and dydx values in xsub, ysub, and dsub respectively.

This does not update any of the values stored in component

outputs_and_jacobian!(component::AbstractExplicitComponent, x)

Evaluate the outputs and their derivatives with respect to the inputs.

Return the result and store in component.y and component.dydx

CoupledSystems.outputs_and_jacobian!! — Function

outputs_and_jacobian!!(component::AbstractExplicitComponent, x)

Force (re-)evaluation of the outputs and their derivatives with respect to the inputs.

Return the result and store in component.y and component.dydx

CoupledSystems.outputs_and_jacobian!!! — Function

outputs_and_jacobian!!!(component::AbstractExplicitComponent, x)

Force (re-)evaluation of the outputs and their derivatives with respect to the inputs for the system. Also force (re-)evaluation of the outputs and jacobians of the system's subcomponents.

Return the result and store in component.y and component.dydx

CoupledSystems.residuals — Function

residuals(component::AbstractImplicitComponent)

Return the residuals stored in component

residuals(component::AbstractImplicitComponent, x, y)

Evaluate the residuals of an implicit system component.

This does not update any of the values stored in component

CoupledSystems.residuals! — Function

residuals!(component::AbstractImplicitComponent, r, x, y)

Evaluate the residuals of an implicit system component and store the result in r.

This does not update any of the values stored in component

residuals!(component::ImplicitSystem, r, xsub, ysub, rsub, x, y)

Evaluate the residuals of an implicit system and store the result in y. Store the intermediate results in xsub, ysub, and rsub.

This does not update any of the values stored in component

residuals!(component::AbstractImplicitComponent, x, y)

Evaluate the residuals of an implicit system component.

Return the result and store in component.r

CoupledSystems.residuals!! — Function

residuals!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residuals of an implicit system component.

Return the result and store in component.r

CoupledSystems.residuals!!! — Function

residuals!!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residuals of an implicit system component. Also force (re-)evaluation of the residuals of the system's subcomponents.

Return the result and store in component.r

CoupledSystems.residual_input_jacobian — Function

residual_input_jacobian(component::AbstractImplicitComponent)

Return the jacobian of the residuals with respect to the inputs stored in component.

residual_input_jacobian(component::AbstractImplicitComponent, x, y)

Evaluate the jacobian of the residuals with respect to the inputs,

This does not update any of the values stored in component

CoupledSystems.residual_input_jacobian! — Function

residual_input_jacobian!(component::ImplicitComponent, drdx, x, y)

Evaluate the jacobian of the residuals with respect to the inputs and store the result in drdx.

This does not update any of the values stored in component

residual_input_jacobian!(component::ImplicitSystem, drdx, xsub, ysub, rsub,
    dsub, x, y)

Evaluate the jacobian of the residual with respect to the inputs and store the result in drdx. Store the intermediate subcomponent x, y and drdx values in xsub, ysub, and dsub respectively.

This does not update any of the values stored in component

residual_input_jacobian!(component::AbstractImplicitComponent, x, y)

Evaluate the jacobian of the residuals with respect to the inputs.

Return the result and store in component.drdx

CoupledSystems.residual_input_jacobian!! — Function

residual_input_jacobian!!(component::ImplicitComponent, x, y)

Force (re-)evaluation of the jacobian of the residuals with respect to the inputs.

Return the result and store in component.drdx

CoupledSystems.residual_input_jacobian!!! — Function

residual_input_jacobian!!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the jacobian of the residuals with respect to the inputs. Also force (re-)evaluation of the jacobians of the system's subcomponents.

Return the result and store in component.drdx

CoupledSystems.residual_output_jacobian — Function

residual_output_jacobian(component::AbstractImplicitComponent)

Return the jacobian of the residuals with respect to the outputs stored in component.

residual_output_jacobian(component::AbstractImplicitComponent, x, y)

Evaluate the jacobian of the residuals with respect to the outputs,

This does not update any of the values stored in component

CoupledSystems.residual_output_jacobian! — Function

residual_output_jacobian!(component::AbstractImplicitComponent, drdy, x, y)

Evaluate the jacobian of the residuals with respect to the outputs and store the result in drdy

This does not update any of the values stored in component

residual_output_jacobian!(component::ImplicitSystem, drdy, xsub, ysub, dxsub,
    dysub, x, y)

Evaluate the jacobian of the residuals with respect to the outputs and store the result in drdx. Store the intermediate subcomponent x, y and drdy values in xsub, ysub, and dsub respectively.

This does not update any of the values stored in component

residual_output_jacobian!(component::AbstractImplicitComponent, x, y)

Evaluate the jacobian of the residuals with respect to the outputs.

Return the result and store in component.drdy

CoupledSystems.residual_output_jacobian!! — Function

residual_output_jacobian!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the jacobian of the residuals with respect to the outputs.

Return the result and store in component.drdy

CoupledSystems.residual_output_jacobian!!! — Function

residual_output_jacobian!!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the jacobian of the residuals with respect to the outputs. Also force (re-)evaluation of the jacobians of the system's subcomponents.

Return the result and store in component.drdy

CoupledSystems.residuals_and_input_jacobian — Function

residuals_and_input_jacobian(component::AbstractImplicitComponent)

Return the residual and jacobian of the residuals with respect to the inputs stored in component.

residuals_and_input_jacobian(component::AbstractImplicitComponent, x, y)

Evaluate the residual and jacobian of the residuals with respect to the inputs.

This does not update any of the values stored in component

CoupledSystems.residuals_and_input_jacobian! — Function

residuals_and_input_jacobian!(component::AbstractImplicitComponent, r, drdx, x, y)

Evaluate the residual and jacobian of the residuals with respect to the inputs and store the result in r and drdx.

This does not update any of the values stored in component

residuals_and_input_jacobian!(component::ImplicitSystem, r, drdx, xsub,
    ysub, rsub, dsub, x, y)

Evaluate the residuals and their derivatives with respect to the inputs and store the results in r and drdx. Store the intermediate subcomponent x, y, r, and drdx values in xsub, ysub, rsub, and dsub respectively.

This does not update any of the values stored in component

residuals_and_input_jacobian!(component::AbstractImplicitComponent, x, y)

Evaluate the residual and jacobian of the residuals with respect to the inputs.

Return the result and store in component.r and component.drdx

CoupledSystems.residuals_and_input_jacobian!! — Function

residuals_and_input_jacobian!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residual and jacobian of the residuals with respect to the inputs.

Return the result and store in component.r and component.drdx

CoupledSystems.residuals_and_input_jacobian!!! — Function

residuals_and_input_jacobian!!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residual and jacobian of the residuals with respect to the inputs. Also force (re-)evaluation of each subcomponents residual and jacobian with respect to its inputs.

Return the result and store in component.r and component.drdx

CoupledSystems.residuals_and_output_jacobian — Function

residuals_and_output_jacobian(component::AbstractImplicitComponent)

Return the residuals and jacobian of the residuals with respect to the outputs stored in component.

residuals_and_output_jacobian(component::AbstractImplicitComponent, x, y)

Evaluate the residuals and jacobian of the residuals with respect to the outputs.

This does not update any of the values stored in component

CoupledSystems.residuals_and_output_jacobian! — Function

residuals_and_output_jacobian!(component::AbstractImplicitComponent, r, drdy, x, y)

Evaluate the residuals and jacobian of the residuals with respect to the outputs and store the results in r and drdy

This does not update any of the values stored in component

residuals_and_output_jacobian!(component::ImplicitSystem, r, drdy, xsub,
    ysub, rsub, dxsub, dysub, x, y)

Evaluate the residuals and their derivatives with respect to the outputs and store the results in r and drdy. Store the intermediate subcomponent x, y, r, drdx, and drdy values in xsub, ysub, rsub, dxsub, and dysub respectively.

This does not update any of the values stored in component

residuals_and_output_jacobian!(component::AbstractImplicitComponent, x, y)

Evaluate the residuals and jacobian of the residuals with respect to the outputs.

Return the result and store in component.r and component.drdy

CoupledSystems.residuals_and_output_jacobian!! — Function

residuals_and_output_jacobian!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residuals and jacobian of the residuals with respect to the outputs.

Return the result and store in component.r and component.drdy

CoupledSystems.residuals_and_output_jacobian!!! — Function

residuals_and_output_jacobian!!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residual and jacobian of the residuals with respect to the outputs. Also force (re-)evaluation of each subcomponents residual and jacobian with respect to its outputs.

Return the result and store in component.r and component.drdy

CoupledSystems.residuals_and_jacobians — Function

residuals_and_jacobians(component::AbstractImplicitComponent)

Return the residual, its jacobian with respect to the inputs, and its jacobian with respect to the outputs stored in component.

residuals_and_jacobians(component::AbstractImplicitComponent, x, y)

Evaluate the residual, its jacobian with respect to the inputs, and its jacobian with respect to the outputs.

This does not update any of the values stored in component

CoupledSystems.residuals_and_jacobians! — Function

residuals_and_jacobians!(component::AbstractImplicitComponent, r, drdx, drdy, x, y)

Evaluate the residual, its jacobian with respect to the inputs, and its jacobian with respect to the outputs and store the results in r, drdx, and drdy

This does not update any of the values stored in component

residuals_and_jacobians!(component::ImplicitSystem, r, drdx, drdy, xsub,
    ysub, rsub, dxsub, dysub, x, y)

Evaluate the residual, its jacobian with respect to the inputs, and its jacobian with respect to the outputs and store the results in r, drdx, and drdy Store the intermediate subcomponent x, y, r, drdx, and drdy values in xsub, ysub, rsub, dxsub, and dysub respectively.

This does not update any of the values stored in component

residuals_and_jacobians!(component::AbstractImplicitComponent, x, y)

Evaluate the residual, its jacobian with respect to the inputs, and its jacobian with respect to the outputs.

Return the result and store in component.r, component.drdx, and component.drdy.

CoupledSystems.residuals_and_jacobians!! — Function

residuals_and_jacobians!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residual, its jacobian with respect to the inputs, and its jacobian with respect to the outputs.

Return the result and store in component.r, component.drdx, and component.drdy.

CoupledSystems.residuals_and_jacobians!!! — Function

residuals_and_jacobians!!!(component::AbstractImplicitComponent, x, y)

Force (re-)evaluation of the residual, its jacobian with respect to the inputs, and its jacobian with respect to the outputs. Also force (re-)evaluation of each subcomponents residual and jacobians.

Return the result and store in component.r, component.drdx, and component.drdy.

Convenience Functions

combine
separate

Private API

CoupledSystems.AbstractChainRuleMode — Type

AbstractChainRuleMode

Type representing direction in which to to apply the chain rule when computing the derivatives of an explicit system.

CoupledSystems.AbstractSensitivityMode — Type

AbstractSensitivityMode

Abstract type representing mode by which to apply the analytic sensitivity equations when computing derivatives of an explicit component created by combining an implicit component, output component, and solver.

CoupledSystems.TrustRegion — Type

TrustRegion{T} <: AbstractSolver

Trust region solver, as provided by the NLsolve package. Stores the keyword arguments which are passed to NLsolve in the field kwargs

CoupledSystems.TrustRegion — Method

TrustRegion(; kwargs...)

Trust region solver, as provided by the NLsolve package. Uses the same default arguments as NLsolve.

The relevant keyword arguments and their default values are repeated here for convenience.

Keyword Arguments

xtol = 0.0: Step size tolerance for ending iterations

ftol = 1e-8: Function tolerance for ending iterations

iterations = 1000: Maximum number of iterations

factor = 1.0: Factor for determining initial size of trust region

autoscale = true: Autoscale variables?

linsolve = (x, A, b) -> copyto!(x, A): Linear solver function

Base.axes — Method

axes(var::NamedVar)

Return the axes of a named variable

Base.length — Method

length(var::NamedVar)

Return the length of a named variable

Base.size — Method

size(var)

Return the size of a named variable

CoupledSystems.allocate_jacobian — Function

allocate_jacobian(x, y, sparsity=DensePattern())

Allocates a jacobian matrix given the inputs, outputs, and (optionally) the sparsity structure.

CoupledSystems.analytic_sensitivity_equation — Function

analytic_sensitivity_equation(dfdx, dfdu, drdx, drdu, mode)

Compute the sensitivities of an implicit system using the analytic sensitivity equation in Direct() or Adjoint() mode.

CoupledSystems.create_mutating_solver_cache — Method

create_mutating_solver_cache(component::AbstractImplicitComponent, x, solver::AbstractSolver)

Construct a cache for a nonlinear solver which uses mutating functions. x is a pointer to an array that will be modified to contain the correct input values.

CoupledSystems.create_nonmutating_solver_cache — Method

create_nonmutating_solver_cache(component::AbstractImplicitComponent, x, solver::AbstractSolver)

Construct a cache for a nonlinear solver which uses nonmutating functions. x is a pointer to an array that will be modified to contain the correct input values.

CoupledSystems.create_output_jacobian_functions — Function

create_output_jacobian_functions(f!, x0, y0, dtype, sp)

Creates functions to compute the jacobian using the specified derivative type and sparsity pattern.

Arguments

f!: In-place version of the output function: y = f(x)

x0::AbstractVector: Initial values for inputs (used for size and type information)

y0::AbstractVector: Initial values for outputs (used for size and type information)

dtype::AbstractDiffMethod: Method used to calculate derivatives

sparsity::AbstractSparsityPattern: Jacobian sparsity structure

CoupledSystems.create_residual_input_jacobian_functions — Function

create_residual_input_jacobian_functions(f!, x0, y0, r0, dtype, sp)

Creates functions to compute the jacobian of the residual with respect to the inputs using the specified derivative type and sparsity pattern.

Arguments

f!: In-place residual function f(r, x, y).

x0: Initial values for inputs (used for size and type information)

y0: Initial values for outputs (used for size and type information)

r0: Initial values for residuals (used for size and type information)

dtype: Method to use to calculate derivatives

sparsity: Jacobian sparsity structure

CoupledSystems.create_residual_output_jacobian_functions — Function

create_residual_output_jacobian_functions(f!, x0, y0, r0, dtype, sp)

Creates functions to compute the jacobian of the residual with respect to the inputs using the specified derivative type and sparsity pattern.

Arguments

f!: In-place residual function f(r, x, y).

x0: Initial values for inputs (used for size and type information)

y0: Initial values for outputs (used for size and type information)

r0: Initial values for residuals (used for size and type information)

dtype: Method to use to calculate derivatives

sparsity: Jacobian sparsity structure

CoupledSystems.finitediff_type — Function

finitediff_type(mode::AbstractFD)

Converts the types used by this package to the inputs expected by the FiniteDiff package.

CoupledSystems.forward_mode_jacobian! — Function

forward_mode_jacobian!(component::ExplicitSystem, dydx, dsub=nothing)

Fills in the system jacobian matrix dydx using the chain rule in forward mode.

If dsub is not provided, jacobian matrices will be taken from each subcomponent's internal storage.

CoupledSystems.forward_mode_jacobian_branch! — Function

forward_mode_jacobian_branch!(system, dydx, ix, cprod, jc, jx, dsub=nothing)

Adds to the jacobian matrix using the chain rule in forward mode.

This function is called recursively to create new branches as necessary.

If dsub is not provided, jacobian matrices will be taken from each subcomponent's internal storage.

CoupledSystems.forward_mode_mapping — Method

forward_mode_mapping(inputs, components, component_input_mapping, output_mapping)

Returns (input_mapping, component_output_mapping) where input_mapping contains a mapping from each system input to component inputs and/or system outputs and reversed_component_mapping contains a mapping from each component output to corresponding component inputs and/or system outputs.

CoupledSystems.getindices — Method

getindices(collection, indices::Tuple)

Custom version of getindex that allow collections to be accessed using a tuple of indices

CoupledSystems.getsparsity — Method

getsparsity(::DensePattern, nx, nf)

Get rows and cols for dense jacobian

CoupledSystems.getsparsity — Method

getsparsity(sp::SparsePattern, nx, nf)

Get rows and cols for sparse jacobian

CoupledSystems.inputs — Method

inputs(component::AbstractComponent)

Return the inputs corresponding to the outputs stored in component

CoupledSystems.output_mapping_matrices — Method

output_mapping_matrices(argin, argstate, argout)

Construct matrices that maps an implicit function's inputs and/or state variables to specified output variables.

CoupledSystems.reverse_mode_jacobian! — Function

reverse_mode_jacobian!(component::ExplicitSystem, dydx, dsub=nothing)

Fills in the system jacobian matrix dydx using the chain rule in reverse mode.

If dsub is not provided, jacobian matrices will be taken from each subcomponent's internal storage.

CoupledSystems.reverse_mode_jacobian_branch! — Function

reverse_mode_jacobian_branch!(component::ExplicitSystem, dydx, iy, cprod, jc,
    jy, dsub=nothing)

Adds to the jacobian matrix using the chain rule in reverse mode.

This function is called recursively to create new branches as necessary.

If dsub is not provided, jacobian matrices will be taken from each subcomponent's internal storage.

CoupledSystems.same_axes — Function

same_axes(v1, v2)

Return v2 with axes correponding to v1. If v1 is a scalar, return a scalar.

CoupledSystems.setindices — Method

setindices(x, v::Tuple, indices::Tuple)

Creates a new collection similar to x with the values at the indices in indices set to the values in v.

CoupledSystems.subcomponent_input_jacobians!! — Method

subcomponent_input_jacobians!!(component::ImplicitSystem, x, y)

Force (re-)evaluation the input jacobians of an implicit system's subcomponents.

Store in the x_dfdx, y_dfdx, and drdx fields of each subcomponent.

CoupledSystems.subcomponent_input_jacobians! — Method

subcomponent_input_jacobians!(component::ImplicitSystem, xsub, ysub, dsub, x, y)

Evaluate the input jacobians of an implicit system's subcomponents. Store the results in xsub, ysub, and dsub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_input_jacobians! — Method

subcomponent_input_jacobians!(component::ImplicitSystem, x, y)

Evaluate the input jacobians of an implicit system's subcomponents.

Store in the x_dfdx, y_dfdx, and drdx fields of each subcomponent.

CoupledSystems.subcomponent_output_jacobians!! — Method

subcomponent_output_jacobians!!(component::ImplicitSystem, x, y)

Force (re-)evaluation the output jacobians of an implicit system's subcomponents.

Store in the x_dfdy, y_dfdy, and drdy fields of each subcomponent.

CoupledSystems.subcomponent_output_jacobians! — Method

subcomponent_output_jacobians!(component::ImplicitSystem, xsub, ysub, dsub, x, y)

Evaluate the output jacobians of an implicit system's subcomponents. Store the results in xsub, ysub, and dsub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_output_jacobians! — Method

subcomponent_output_jacobians!(component::ImplicitSystem, x, y)

Evaluate the output jacobians of an implicit system's subcomponents.

Store in the x_dfdy, y_dfdy, and drdy fields of each subcomponent.

CoupledSystems.subcomponent_outputs!! — Method

subcomponent_outputs!!(component::ExplicitSystem, x)

Force re-evaluation of the inputs and outputs of an explicit system's subcomponents.

Store in the x_f and y fields of each subcomponent.

CoupledSystems.subcomponent_outputs! — Method

subcomponent_outputs!(component::ExplicitSystem, xsub, ysub, x)

Evaluate the inputs and outputs of an explicit systems subcomponents and store the results in xsub and ysub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_outputs! — Method

subcomponent_outputs!(component::ExplicitSystem, x)

Evaluate the inputs and outputs of an explicit system's subcomponents.

Store in the x_f and y fields of each subcomponent.

CoupledSystems.subcomponent_outputs_and_jacobians!! — Method

subcomponent_outputs_and_jacobians!!(component::ExplicitSystem, x)

Force (re-)evaluation of inputs, outputs, and jacobians of an explicit system's subcomponents.

Store in the x_f, y, and dydx fields of each subcomponent.

CoupledSystems.subcomponent_outputs_and_jacobians! — Method

subcomponent_outputs_and_jacobians!(component::ExplicitSystem, xsub, ysub, dsub, x)

Evaluate the inputs, outputs, and jacobians of an explicit system's subcomponents. Store the results in xsub, ysub, and dsub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_outputs_and_jacobians! — Method

subcomponent_outputs_and_jacobians!(component::ExplicitSystem, x)

Evaluate the inputs, outputs, and jacobians of an explicit system's subcomponents.

Store in the x_f, y, and dydx fields of each subcomponent.

CoupledSystems.subcomponent_residuals!! — Method

subcomponent_residuals!!(component::ImplicitSystem, x, y)

Force re-evaluation of the residuals of an explicit system's subcomponents.

Store in the x_f, y_f, and r fields of each subcomponent.

CoupledSystems.subcomponent_residuals! — Method

subcomponent_residuals!(component::ImplicitSystem, xsub, ysub, rsub, x, y)

Evaluate the residuals of an implicit system's subcomponents and store the results in xsub, ysub, and rsub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_residuals! — Method

subcomponent_residuals!(component::ImplicitSystem, x, y)

Evaluate the residuals of an implicit system's subcomponents.

Store in the x_f, y_f, and r fields of each subcomponent.

CoupledSystems.subcomponent_residuals_and_input_jacobians!! — Method

subcomponent_residuals_and_input_jacobians!!(component::ImplicitSystem, x, y)

Force (re-)evaluation the residuals and input jacobians of an implicit system's subcomponents.

Store in the x_dfdx, y_dfdx, r, and drdx fields of each subcomponent.

CoupledSystems.subcomponent_residuals_and_input_jacobians! — Method

subcomponent_residuals_and_input_jacobians!(component::ImplicitSystem, xsub,
    ysub, rsub, dsub, x, y)

Evaluate the residuals and input jacobians of an implicit system's subcomponents. Store the results in xsub, ysub, rsub, and dsub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_residuals_and_input_jacobians! — Method

subcomponent_residuals_and_input_jacobians!(component::ImplicitSystem, x, y)

Evaluate the residuals and input jacobians of an implicit system's subcomponents.

Store in the x_dfdx, y_dfdx, r, and drdx fields of each subcomponent.

CoupledSystems.subcomponent_residuals_and_jacobians!! — Method

subcomponent_residuals_and_jacobians!!(component::ImplicitSystem, x, y)

Force (re-)evaluation of the residuals and output jacobians of an implicit system's subcomponents.

Store in the x_dfdy, y_dfdy, r, and drdy fields of each subcomponent.

CoupledSystems.subcomponent_residuals_and_jacobians! — Method

subcomponent_residuals_and_jacobians!(component::ImplicitSystem, xsub,
    ysub, rsub, dsub, x, y)

Evaluate the residuals and jacobians of an implicit system's subcomponents. Store the results in xsub, ysub, rsub, dxsub, and dysub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_residuals_and_jacobians! — Method

subcomponent_residuals_and_jacobians!(component::ImplicitSystem, x, y)

Evaluate the residuals and jacobians of an implicit system's subcomponents.

Store in the x_dfdy, y_dfdy, r, drdx, and drdy fields of each subcomponent.

CoupledSystems.subcomponent_residuals_and_output_jacobians!! — Method

subcomponent_residuals_and_output_jacobians!!(component::ImplicitSystem, x, y)

Force (re-)evaluation of the residuals and output jacobians of an implicit system's subcomponents.

Store in the x_dfdy, y_dfdy, r, and drdy fields of each subcomponent.

CoupledSystems.subcomponent_residuals_and_output_jacobians! — Method

subcomponent_residuals_and_output_jacobians!(component::ImplicitSystem, xsub,
    ysub, rsub, dsub, x, y)

Evaluate the residuals and output jacobians of an implicit system's subcomponents. Store the results in xsub, ysub, rsub, and dsub.

This does not update any of the values stored in component

CoupledSystems.subcomponent_residuals_and_output_jacobians! — Method

subcomponent_residuals_and_output_jacobians!(component::ImplicitSystem, x, y)

Evaluate the residuals and output jacobians of an implicit system's subcomponents.

Store in the x_dfdy, y_dfdy, r, and drdy fields of each subcomponent.

CoupledSystems.system_component_mapping — Method

system_component_mapping(components, argin)

Constructs the component input-output mapping for a system based on input and output argument names.

CoupledSystems.system_output_mapping — Method

system_output_mapping(components, argin, argout)

Constructs the output mapping for a system based on input and output argument names of the system and components.

CoupledSystems.update_system_input_jacobian! — Method

update_system_input_jacobian!(system::ImplicitSystem, drdx, dsub)

Update the input jacobian drdx to correspond to the subcomponent input jacobians in dsub.

CoupledSystems.update_system_input_jacobian! — Method

update_system_input_jacobian!(system::ImplicitSystem)

Update the input jacobian matrix in component to correspond to the subcomponent residuals.

CoupledSystems.update_system_output_jacobian! — Method

update_system_output_jacobian!(system::ImplicitSystem, drdy, dsub)

Update the output jacobian drdy to correspond to the subcomponent output jacobians in dsub.

CoupledSystems.update_system_output_jacobian! — Method

update_system_output_jacobian!(system::ImplicitSystem)

Update the output jacobian in system to correspond to the subcomponent input and output jacobians.

CoupledSystems.update_system_outputs! — Method

update_system_outputs!(component, y, x, ysub)

Update the outputs y given the inputs x and the component outputs ysub.

CoupledSystems.update_system_outputs! — Method

update_system_outputs!(component, x)

Update the outputs in component given the inputs x.

CoupledSystems.update_system_residuals! — Method

update_system_residuals!(component, r, rsub)

Update the residuals r to make them correspond to the subcomponent residuals in rsub.

CoupledSystems.update_system_residuals! — Method

update_system_residuals!(component)

Update the residuals in component to correspond to the subcomponent residuals.

CoupledSystems.vectorize — Function

vectorize(x)

Converts x to a vector.

@index ```